02_real_estate_roi_calculator/requirements.txt
appnope==0.1.2
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
boto3==1.17.10
botocore==1.20.10
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
confuse==1.4.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
googlemaps==4.4.2
htmlmin==0.1.12
idna==2.10
ImageHash==4.2.0
ipykernel==5.4.3
ipython==7.20.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
llvmlite==0.35.0
MarkupSafe==1.1.1
matplotlib==3.3.4
missingno==0.4.2
mistune==0.8.4
nbclient==0.5.2
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
networkx==2.5
notebook==6.2.0
numba==0.52.0
numpy==1.20.1
packaging==20.9
pandas==1.2.2
pandas-profiling==2.11.0
pandocfilters==1.4.3
parso==0.8.1
pexpect==4.8.0
phik==0.11.0
pickleshare==0.7.5
Pillow==8.1.0
prometheus-client==0.9.0
prompt-toolkit==3.0.16
psycopg2==2.8.6
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.8.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.25.1
s3transfer==0.3.4
scipy==1.6.1
seaborn==0.11.1
Send2Trash==1.5.0
six==1.15.0
tangled-up-in-unicode==0.0.6
terminado==0.9.2
testpath==0.4.4
tornado==6.1
tqdm==4.57.0
traitlets==5.0.5
urllib3==1.26.3
visions==0.6.0
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
